requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
